Add ast::CallExpr and its parser and printer
ast::CallExpr is just a function call expression:

For example if we have previously defined a function
called 'f', we can do something like 'f(3)' -
that is call the function 'f' with the first argument
being the number literal 3.

Note: This can't be used as a statement (like in C),
you always have to do something like 'variable = f(3)',
whereas the statement version is often used for procedures,
something like 'do_stuff();'.

This commit was made in the air! :)